Improved performance when getting market price
Made getmarketprices do all the checks (previously was in
getmarketprice) and return a dictionary ready to go. This way,
getmarketprice only has to lookup the item index and so when called for
many items, there are less loops to go through. This improved the warmup
time for the app if data wasn't in the cache.

Added getpaintimages.py - script for getting paint can images from wiki.
Added test.py - to do some time tests

getitems in tf2api.py and getitemsdict in tf2search.py now return an
OrderedDict instead of a regular dictionary.